#15: Each entity has its own router and model
