+++ backend/src/models/accountTransaction.ts
@@ -1,7 +1,66 @@
-import { OkPacket } from 'mysql2';
+import { OkPacket, RowDataPacket } from 'mysql2';
 import { mysql as sql } from 'yesql';
 import { db } from '../db';
-import { AccountTransaction } from '../types/account';
+import { AccountTransaction, AccountTransactionSummary, rowToAccountTransaction, rowToAccountTransactionSummary } from '../types/account';
+
+export interface AccountTransactionParams {
+  userId: number,
+  accountId?: number,
+  type?: string,
+}
+
+export const findOne = (userId: number, id: number): Promise<AccountTransaction> => {
+  let queryString = `
+    SELECT *
+    FROM account_transaction_detailed AS atd
+    WHERE
+      atd.user_id = :userId
+      AND atd.id = :id
+  `;
+
+  return new Promise((resolve, reject) => {
+    db.query(
+      sql(queryString)({ userId, id }),
+      (err, result) => {
+        if (err) { reject(err); return; }
+
+        const row = (<RowDataPacket>result)[0];
+        resolve(rowToAccountTransaction(row));
+      },
+    );
+  });
+};
+
+
+export const findAll = (params: AccountTransactionParams): Promise<AccountTransactionSummary[]> => {
+  let queryString = `
+    SELECT *
+    FROM account_transaction_summary AS ats
+  `;
+
+  const filters = ['ats.user_id = :userId'];
+  if (params.accountId) {
+    filters.push('ats.account_id = :accountId');
+  }
+  if (params.type) {
+    filters.push('ats.type = :type');
+  }
+  queryString += `
+  WHERE ${filters.join(' AND ')}
+  `;
+
+  return new Promise((resolve, reject) => {
+    db.query(
+      sql(queryString)({ ...params }),
+      (err, result) => {
+        if (err) { reject(err); return; }
+
+        const rows = <RowDataPacket[]>result;
+        resolve(rows.map(rowToAccountTransactionSummary));
+      },
+    );
+  });
+};
 
 export const doesExist = (accountTransaction: AccountTransaction): Promise<boolean> => {
   const getComparison = (propertyName: string) => {
