+++ backend/src/routes/portfolioRouter.ts
@@ -1,10 +1,10 @@
 import express, { Request, Response } from 'express';
-import * as securityHistoryModel from '../../models/securityHistory';
-import { handleRequest } from '../../utils/server';
+import * as securityHistoryModel from '../models/securityPrice';
+import { handleRequest } from '../utils/server';
 
-const portfolioHistoryRouter = express.Router();
+const portfolioRouter = express.Router();
 
-portfolioHistoryRouter.get('/:currency', async (req: Request, res: Response) => {
+portfolioRouter.get('/history/:currency', async (req: Request, res: Response) => {
   const currency = String(req.params.currency);
   const startDate = new Date(String(req.query.start));
   const endDate = new Date(String(req.query.end));
@@ -13,4 +13,4 @@ portfolioHistoryRouter.get('/:currency', async (req: Request, res: Response) => 
   handleRequest(res, securityHistoryModel.getPortfolioHistory(userId, currency, startDate, endDate));
 });
 
-export { portfolioHistoryRouter };
+export { portfolioRouter };
