+++ backend/src/types/security.ts
@@ -1,4 +1,6 @@
-export type SecurityTransactionType = 'buy' | 'sell' | 'posting' | 'dividend' | 'tax_refund';
+import { RowDataPacket } from 'mysql2';
+
+export type SecurityTransactionType = 'buy' | 'sell' | 'posting' | 'vesting' | 'dividend' | 'tax_refund';
 
 export interface Security {
   id: number,
@@ -56,6 +58,25 @@ export interface SecurityTransactionSummary extends SecurityTransaction {
   moneyId: number,
   nameShort: string,
 }
+
+export const rowToSecurityHistory = (row: RowDataPacket): SecurityTransactionSummary => ({
+  id: row.id,
+  symbol: row.symbol,
+  nameShort: row.name_short,
+  date: new Date(row.date),
+  type: row.type,
+  accountId: row.account_id,
+  accountTransactionId: row.account_transaction_id,
+  securityId: row.security_id,
+  moneyId: row.money_id,
+  currency: row.currency,
+  price: Number(row.price),
+  amount: Number(row.amount),
+  total: Number(row.total),
+  value: Number(row.value),
+  fee: Number(row.fee),
+  tax: Number(row.tax),
+});
 
 export interface SecurityTransactionForeign extends SecurityTransaction {
   exchangeToValue: number,
